Implement Package installer
